With d9009dc, this fixes an issue that users had when they upgraded and downgraded history-service often. It also prevents the issue from happening again.

The issue was that downgrading history-service (most likely by switching from rc to stable channels) would also cause the schema_version in the history database to be overwritten to the old version. However, the database's schema was unchanged when this "downgrade" happened. So users would be running the v18 schema with a version of history-service that expected (and wrote) v17. This isn't usually a problem because our schema changes are backwards-compatible. However, they are not idempotent: you can't run the v19 schema update multiple times with the same results every time. Instead, since there is no way to add a column only if it doesn't exist, the update errors out. This prevented a possible future update to v20 since v19 has to be applied without errors before v20 happens.

One commit here makes the v19 schema update idempotent. This allows users who have a v19 schema that is masquerading as v18 to upgrade to this version of history-service without issue.

The other commit makes it so history-service won't put a smaller schema_version in the database than already exists. This prevents the same issue from happening in the future. With this done, we won't need to make idempotent schema updates after v19.

+++ plugins/sqlite/sqlitedatabase.cpp
@@ -195,6 +195,22 @@ void trace(void* /* something */, const char *query)
     qDebug() << "SQLITE TRACE:" << query;
 }
 
+
+bool SQLiteDatabase::upgradeNeeded(int version) const
+{
+
+    QSqlQuery query(mDatabase);
+
+    if (version == 19) {
+        // check for already exist column
+        if (query.exec("SELECT sentTime FROM text_events LIMIT 1")) {
+            return false;
+        }
+    }
+
+    return true;
+}
+
 bool SQLiteDatabase::createOrUpdateDatabase()
 {
     bool create = !QFile(mDatabasePath).exists();
@@ -238,7 +254,10 @@ bool SQLiteDatabase::createOrUpdateDatabase()
         existingVersion = query.value(0).toInt();
         upgradeToVersion = existingVersion + 1;
         while (upgradeToVersion <= mSchemaVersion) {
-            statements += parseSchemaFile(QString(":/database/schema/v%1.sql").arg(QString::number(upgradeToVersion)));
+
+            if (upgradeNeeded(upgradeToVersion)) {
+                statements += parseSchemaFile(QString(":/database/schema/v%1.sql").arg(QString::number(upgradeToVersion)));
+            }
             ++upgradeToVersion;
         }
     }
@@ -250,17 +269,20 @@ bool SQLiteDatabase::createOrUpdateDatabase()
         return false;
     }
 
-    // now set the new database schema version
-    if (!query.exec("DELETE FROM schema_version")) {
-        qCritical() << "Failed to remove previous schema versions. SQL Statement:" << query.lastQuery() << "Error:" << query.lastError();
-        rollbackTransaction();
-        return false;
-    }
+    //don't downgrade database version
+    if (mSchemaVersion > existingVersion) {
+        // now set the new database schema version
+        if (!query.exec("DELETE FROM schema_version")) {
+            qCritical() << "Failed to remove previous schema versions. SQL Statement:" << query.lastQuery() << "Error:" << query.lastError();
+            rollbackTransaction();
+            return false;
+        }
 
-    if (!query.exec(QString("INSERT INTO schema_version VALUES (%1)").arg(mSchemaVersion))) {
-        qCritical() << "Failed to insert new schema version. SQL Statement:" << query.lastQuery() << "Error:" << query.lastError();
-        rollbackTransaction();
-        return false;
+        if (!query.exec(QString("INSERT INTO schema_version VALUES (%1)").arg(mSchemaVersion))) {
+            qCritical() << "Failed to insert new schema version. SQL Statement:" << query.lastQuery() << "Error:" << query.lastError();
+            rollbackTransaction();
+            return false;
+        }
     }
 
     // now check if any data updating is required
